Make rock, paper, scisccors conole game
Created a 5 round rock paper scissors game to be played in console. It
has 3 main functions - one to get human choice, the other for computer
choie and the last one to play the round with these choices. They are
all put into another function where these functions are looped 5 times
to play a game.